Rename abstract classes
