+++ airbyte-oauth/src/main/java/io/airbyte/oauth/BaseOAuthFlow.java
@@ -5,236 +5,83 @@
 package io.airbyte.oauth;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.google.common.collect.ImmutableMap;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-import io.airbyte.commons.json.Jsons;
+import io.airbyte.config.ConfigSchema;
+import io.airbyte.config.DestinationOAuthParameter;
+import io.airbyte.config.SourceOAuthParameter;
 import io.airbyte.config.persistence.ConfigNotFoundException;
 import io.airbyte.config.persistence.ConfigRepository;
+import io.airbyte.validation.json.JsonValidationException;
 import java.io.IOException;
-import java.lang.reflect.Type;
-import java.net.URI;
-import java.net.URLEncoder;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
-import java.util.function.Function;
-import java.util.function.Supplier;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-/*
- * Class implementing generic oAuth 2.0 flow.
+/**
+ * Abstract Class implementing common base methods for managing:
+ * - oAuth config (instance-wide) parameters
+ * - oAuth specifications
  */
-public abstract class BaseOAuthFlow extends BaseOAuthConfig {
+public abstract class BaseOAuthFlow implements OAuthFlowImplementation {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(BaseOAuthFlow.class);
+  private final ConfigRepository configRepository;
 
-  /**
-   * Simple enum of content type strings and their respective encoding functions used for POSTing the
-   * access token request
-   */
-  public enum TOKEN_REQUEST_CONTENT_TYPE {
-
-    URL_ENCODED("application/x-www-form-urlencoded", BaseOAuthFlow::toUrlEncodedString),
-    JSON("application/json", BaseOAuthFlow::toJson);
-
-    String contentType;
-    Function<Map<String, String>, String> converter;
-
-    TOKEN_REQUEST_CONTENT_TYPE(final String contentType, final Function<Map<String, String>, String> converter) {
-      this.contentType = contentType;
-      this.converter = converter;
-    }
-
+  public BaseOAuthFlow(final ConfigRepository configRepository) {
+    this.configRepository = configRepository;
   }
 
-  private final TOKEN_REQUEST_CONTENT_TYPE tokenReqContentType;
-  protected HttpClient httpClient;
-  private final Supplier<String> stateSupplier;
-
-  public BaseOAuthFlow(final ConfigRepository configRepository, final HttpClient httpClient) {
-    this(configRepository, httpClient, BaseOAuthFlow::generateRandomState);
-  }
-
-  public BaseOAuthFlow(final ConfigRepository configRepository, final HttpClient httpClient, final Supplier<String> stateSupplier) {
-    this(configRepository, httpClient, stateSupplier, TOKEN_REQUEST_CONTENT_TYPE.URL_ENCODED);
-  }
-
-  public BaseOAuthFlow(final ConfigRepository configRepository,
-                       final HttpClient httpClient,
-                       final Supplier<String> stateSupplier,
-                       final TOKEN_REQUEST_CONTENT_TYPE tokenReqContentType) {
-    super(configRepository);
-    this.httpClient = httpClient;
-    this.stateSupplier = stateSupplier;
-    this.tokenReqContentType = tokenReqContentType;
-  }
-
-  @Override
-  public String getSourceConsentUrl(final UUID workspaceId, final UUID sourceDefinitionId, final String redirectUrl)
-      throws IOException, ConfigNotFoundException {
-    final JsonNode oAuthParamConfig = getSourceOAuthParamConfig(workspaceId, sourceDefinitionId);
-    return formatConsentUrl(sourceDefinitionId, getClientIdUnsafe(oAuthParamConfig), redirectUrl);
-  }
-
-  @Override
-  public String getDestinationConsentUrl(final UUID workspaceId, final UUID destinationDefinitionId, final String redirectUrl)
-      throws IOException, ConfigNotFoundException {
-    final JsonNode oAuthParamConfig = getDestinationOAuthParamConfig(workspaceId, destinationDefinitionId);
-    return formatConsentUrl(destinationDefinitionId, getClientIdUnsafe(oAuthParamConfig), redirectUrl);
-  }
-
-  /**
-   * Depending on the OAuth flow implementation, the URL to grant user's consent may differ,
-   * especially in the query parameters to be provided. This function should generate such consent URL
-   * accordingly.
-   */
-  protected abstract String formatConsentUrl(UUID definitionId, String clientId, String redirectUrl) throws IOException;
-
-  private static String generateRandomState() {
-    return RandomStringUtils.randomAlphanumeric(7);
-  }
-
-  /**
-   * Generate a string to use as state in the OAuth process.
-   */
-  protected String getState() {
-    return stateSupplier.get();
-  }
-
-  @Override
-  public Map<String, Object> completeSourceOAuth(
-                                                 final UUID workspaceId,
-                                                 final UUID sourceDefinitionId,
-                                                 final Map<String, Object> queryParams,
-                                                 final String redirectUrl)
-      throws IOException, ConfigNotFoundException {
-    final JsonNode oAuthParamConfig = getSourceOAuthParamConfig(workspaceId, sourceDefinitionId);
-    return completeOAuthFlow(
-        getClientIdUnsafe(oAuthParamConfig),
-        getClientSecretUnsafe(oAuthParamConfig),
-        extractCodeParameter(queryParams),
-        redirectUrl,
-        oAuthParamConfig);
-  }
-
-  @Override
-  public Map<String, Object> completeDestinationOAuth(final UUID workspaceId,
-                                                      final UUID destinationDefinitionId,
-                                                      final Map<String, Object> queryParams,
-                                                      final String redirectUrl)
-      throws IOException, ConfigNotFoundException {
-    final JsonNode oAuthParamConfig = getDestinationOAuthParamConfig(workspaceId, destinationDefinitionId);
-    return completeOAuthFlow(
-        getClientIdUnsafe(oAuthParamConfig),
-        getClientSecretUnsafe(oAuthParamConfig),
-        extractCodeParameter(queryParams),
-        redirectUrl, oAuthParamConfig);
-  }
-
-  protected Map<String, Object> completeOAuthFlow(final String clientId,
-                                                  final String clientSecret,
-                                                  final String authCode,
-                                                  final String redirectUrl,
-                                                  final JsonNode oAuthParamConfig)
-      throws IOException {
-    final var accessTokenUrl = getAccessTokenUrl();
-    final HttpRequest request = HttpRequest.newBuilder()
-        .POST(HttpRequest.BodyPublishers
-            .ofString(tokenReqContentType.converter.apply(getAccessTokenQueryParameters(clientId, clientSecret, authCode, redirectUrl))))
-        .uri(URI.create(accessTokenUrl))
-        .header("Content-Type", tokenReqContentType.contentType)
-        .header("Accept", "application/json")
-        .build();
+  protected JsonNode getSourceOAuthParamConfig(final UUID workspaceId, final UUID sourceDefinitionId) throws IOException, ConfigNotFoundException {
     try {
-      HttpResponse<String> response;
-      response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
-      return extractRefreshToken(Jsons.deserialize(response.body()), accessTokenUrl);
-    } catch (final InterruptedException e) {
-      throw new IOException("Failed to complete OAuth flow", e);
-    }
-  }
-
-  /**
-   * Query parameters to provide the access token url with.
-   */
-  protected Map<String, String> getAccessTokenQueryParameters(final String clientId,
-                                                              final String clientSecret,
-                                                              final String authCode,
-                                                              final String redirectUrl) {
-    return ImmutableMap.<String, String>builder()
-        // required
-        .put("client_id", clientId)
-        .put("redirect_uri", redirectUrl)
-        .put("client_secret", clientSecret)
-        .put("code", authCode)
-        .build();
-  }
-
-  /**
-   * Once the user is redirected after getting their consent, the API should redirect them to a
-   * specific redirection URL along with query parameters. This function should parse and extract the
-   * code from these query parameters in order to continue the OAuth Flow.
-   */
-  protected String extractCodeParameter(final Map<String, Object> queryParams) throws IOException {
-    if (queryParams.containsKey("code")) {
-      return (String) queryParams.get("code");
-    } else {
-      throw new IOException("Undefined 'code' from consent redirected url.");
-    }
-  }
-
-  /**
-   * Returns the URL where to retrieve the access token from.
-   */
-  protected abstract String getAccessTokenUrl();
-
-  protected Map<String, Object> extractRefreshToken(final JsonNode data, final String accessTokenUrl) throws IOException {
-    final Map<String, Object> result = new HashMap<>();
-    if (data.has("refresh_token")) {
-      result.put("refresh_token", data.get("refresh_token").asText());
-    } else if (data.has("access_token")) {
-      result.put("access_token", data.get("access_token").asText());
-    } else {
-      LOGGER.info("Oauth flow failed. Data received from server: {}", data);
-      throw new IOException(String.format("Missing 'refresh_token' in query params from %s. Response: %s", accessTokenUrl));
-
-    }
-    return Map.of("credentials", result);
-
-  }
-
-  private static String urlEncode(final String s) {
-    try {
-      return URLEncoder.encode(s, StandardCharsets.UTF_8);
-    } catch (final Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private static String toUrlEncodedString(final Map<String, String> body) {
-    final StringBuilder result = new StringBuilder();
-    for (final var entry : body.entrySet()) {
-      if (result.length() > 0) {
-        result.append("&");
+      final Optional<SourceOAuthParameter> param = MoreOAuthParameters.getSourceOAuthParameter(
+          configRepository.listSourceOAuthParam().stream(), workspaceId, sourceDefinitionId);
+      if (param.isPresent()) {
+        return param.get().getConfiguration();
+      } else {
+        throw new ConfigNotFoundException(ConfigSchema.SOURCE_OAUTH_PARAM, "Undefined OAuth Parameter.");
       }
-      result.append(entry.getKey()).append("=").append(urlEncode(entry.getValue()));
+    } catch (final JsonValidationException e) {
+      throw new IOException("Failed to load OAuth Parameters", e);
     }
-    return result.toString();
   }
 
-  protected static String toJson(final Map<String, String> body) {
-    final Gson gson = new Gson();
-    final Type gsonType = new TypeToken<Map<String, String>>() {
-
-    }.getType();
-    return gson.toJson(body, gsonType);
+  protected JsonNode getDestinationOAuthParamConfig(final UUID workspaceId, final UUID destinationDefinitionId)
+      throws IOException, ConfigNotFoundException {
+    try {
+      final Optional<DestinationOAuthParameter> param = MoreOAuthParameters.getDestinationOAuthParameter(
+          configRepository.listDestinationOAuthParam().stream(), workspaceId, destinationDefinitionId);
+      if (param.isPresent()) {
+        return param.get().getConfiguration();
+      } else {
+        throw new ConfigNotFoundException(ConfigSchema.DESTINATION_OAUTH_PARAM, "Undefined OAuth Parameter.");
+      }
+    } catch (final JsonValidationException e) {
+      throw new IOException("Failed to load OAuth Parameters", e);
+    }
   }
 
+  /**
+   * Throws an exception if the client ID cannot be extracted. Subclasses should override this to
+   * parse the config differently.
+   *
+   * @return The configured Client ID used for this oauth flow
+   */
+  protected String getClientIdUnsafe(final JsonNode oauthConfig) {
+    if (oauthConfig.get("client_id") != null) {
+      return oauthConfig.get("client_id").asText();
+    } else {
+      throw new IllegalArgumentException("Undefined parameter 'client_id' necessary for the OAuth Flow.");
+    }
+  }
+
+  /**
+   * Throws an exception if the client secret cannot be extracted. Subclasses should override this to
+   * parse the config differently.
+   *
+   * @return The configured client secret for this OAuthFlow
+   */
+  protected String getClientSecretUnsafe(final JsonNode oauthConfig) {
+    if (oauthConfig.get("client_secret") != null) {
+      return oauthConfig.get("client_secret").asText();
+    } else {
+      throw new IllegalArgumentException("Undefined parameter 'client_secret' necessary for the OAuth Flow.");
+    }
+  }
 }
