+++ airbyte-oauth/src/main/java/io/airbyte/oauth/flows/SalesforceOAuthFlow.java
@@ -9,7 +9,7 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.config.persistence.ConfigRepository;
-import io.airbyte.oauth.BaseOAuthFlow;
+import io.airbyte.oauth.BaseOAuth2Flow;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.net.http.HttpClient;
@@ -22,7 +22,7 @@ import org.apache.http.client.utils.URIBuilder;
  * Following docs from
  * https://help.salesforce.com/s/articleView?language=en_US&amp;id=sf.remoteaccess_oauth_web_server_flow.htm
  */
-public class SalesforceOAuthFlow extends BaseOAuthFlow {
+public class SalesforceOAuthFlow extends BaseOAuth2Flow {
   // Clickable link for IDE
   // https://help.salesforce.com/s/articleView?language=en_US&id=sf.remoteaccess_oauth_web_server_flow.htm
