+++ airbyte-oauth/src/main/java/io/airbyte/oauth/flows/google/GoogleOAuthFlow.java
@@ -7,7 +7,7 @@ package io.airbyte.oauth.flows.google;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import io.airbyte.config.persistence.ConfigRepository;
-import io.airbyte.oauth.BaseOAuthFlow;
+import io.airbyte.oauth.BaseOAuth2Flow;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.net.http.HttpClient;
@@ -19,7 +19,7 @@ import org.apache.http.client.utils.URIBuilder;
 /**
  * Following docs from https://developers.google.com/identity/protocols/oauth2/web-server
  */
-public abstract class GoogleOAuthFlow extends BaseOAuthFlow {
+public abstract class GoogleOAuthFlow extends BaseOAuth2Flow {
 
   private static final String ACCESS_TOKEN_URL = "https://oauth2.googleapis.com/token";
